fix linting (#1300)
* fix linting

It turns out that in order to support the "import" syntax we need to use
the babel-linter. babel recommends "import" but actually that is not
standard javascript yet, and because it is not standard, estlint won't
support it. So basically so long as we want to use it we need to force
js standard to use babel-eslint instead of eslint.

* fix linting project-wide

We weren't actually using the linter in a lot of places. I've fixed all
of them by adding the linter and fixing all linting issues.

The new strategy works like this:
- use js standard with babel-eslint
- make sure js standard and babel-eslint are devDependencies